+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+## [v0.1.0] - 2021-12-40
+### Changed
+- `separate_rows` and `separate_columns` now contain the row and column indices
+  *before* which horizontal/vertical lines should be drawn (used to be *after* which).
+
 ## [v0.0.3] - 2021-09-06
 ### Added
 - keywords `prefix` and `suffix` to make jupyter output more customizable
@@ -29,7 +34,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 - initialized
 
-[Unreleased]: https://github.com/benmaier/epipack/compare/v0.0.3...HEAD
+[Unreleased]: https://github.com/benmaier/epipack/compare/v0.1.0...HEAD
+[v0.0.3]: https://github.com/benmaier/epipack/compare/v0.0.3...v0.1.0]
 [v0.0.3]: https://github.com/benmaier/epipack/compare/v0.0.2...v0.0.3]
 [v0.0.2]: https://github.com/benmaier/epipack/compare/v0.0.1...v0.0.2]
 [v0.0.1]: https://github.com/benmaier/epipack/compare/v0.0.0...v0.0.1]